feat(django): use gunnicorn ass web server
Previously, the default Django web server was
used but it's not a robust solution.

Gunicorn 'Green Unicorn' is a Python WSGI HTTP Server
for UNIX. It's a pre-fork worker model. The Gunicorn
server is broadly compatible with various web frameworks,
simply implemented, light on server resources, and fairly speedy.

This commit install gunicorn in the hello-django image.